🛠️ :: companyId int type -> long type 변경

+++ data/src/main/java/team/retum/data/remote/api/CompanyApi.kt
@@ -17,7 +17,7 @@ interface CompanyApi {
 
     @GET(JobisUrl.Company.details)
     suspend fun fetchCompanyDetails(
-        @Path("company-id") companyId: Int,
+        @Path("company-id") companyId: Long,
     ): FetchCompanyDetailResponse
 
     @GET(JobisUrl.Company.reviewableCompanies)
