+++ presentation/src/main/java/team/retum/jobis_android/contract/company/CompanyContract.kt
@@ -7,16 +7,16 @@ import team.retum.jobis_android.util.mvi.SideEffect
 import team.retum.jobis_android.util.mvi.State
 import java.util.Collections.emptyList
 
-sealed class CompanySideEffect: SideEffect{
-    object NotFoundCompany: CompanySideEffect()
-    class Exception(val message: String): CompanySideEffect()
+sealed class CompanySideEffect : SideEffect {
+    object NotFoundCompany : CompanySideEffect()
+    class Exception(val message: String) : CompanySideEffect()
 }
 
 data class CompanyState(
     var companies: List<CompanyEntity> = mutableListOf(),
     var page: Int = 1,
     var name: String? = null,
-    var companyId: Int = 0,
+    var companyId: Long = 0,
     val reviewableCompanies: List<ReviewableCompanyEntity> = emptyList(),
     var companyDetails: CompanyDetailsEntity = CompanyDetailsEntity(
         address1 = "",
@@ -39,4 +39,4 @@ data class CompanyState(
         workerNumber = 0,
         zipCode1 = "",
     ),
-): State
+) : State
